+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Use direct dependency versions in pom to simplify resolution (dependencyManagement section remains but for usage as BOM only)
 * Fix lambda modules support (modules declared with lambda expression) (#160)
+* Exclude "sun.*" objects from configuration analysis (#170, #180)
 
 Lambda modules reporting:
 - Root lambda module class will be shown in the diagnostic report (in a list of root modules)